lesson6/app/requirements.txt
chromadb==0.5.11
pyTelegramBotAPI==4.23.0
python-dotenv
watchdog
psutil
aiohttp
pandas
loguru
dependency_injector
sqlmodel
sqlalchemy
psycopg
psycopg-binary